refactor (lexer) : Split code in files
