+++ include/lexer.h
@@ -6,7 +6,7 @@
 /*   By: amennad <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 14:49:53 by amennad           #+#    #+#             */
-/*   Updated: 2023/10/17 17:43:23 by amennad          ###   ########.fr       */
+/*   Updated: 2023/10/18 14:17:10 by amennad          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,9 @@ void	lexer_push_new_node(t_msh *msh, t_lexer_list	*new_node);
 void	lexer_push(t_msh *msh, char *str, t_lexer_type type);
 
 //check cara
+int		is_blank(t_msh *msh, char *prompt, int *i);
 int		is_pipe(t_msh *msh, char *prompt, int *i);
+int		is_right_bracket(t_msh *msh, char *prompt, int *i);
+int		is_left_bracket(t_msh *msh, char *prompt, int *i);
 
 #endif
